// src/app/models/annonce.ts
export interface Annonce {
  id: string; // Identifiant de l'annonce (chaîne de caractères)
  title: string; // Titre de l'annonce (chaîne de caractères)
  description: string; // Description de l'annonce (chaîne de caractères)
  categoryId: string; // Identifiant de la catégorie à laquelle l'annonce est associée (chaîne de caractères)
  userId: string; // Identifiant de l'utilisateur associé à l'annonce (chaîne de caractères)
}